+++ js/mgmIngredient.js
@@ -216,6 +216,11 @@ $('#cloneIng').on('click', '[id*=cloneME]', function () {
 	});
 });
 
+
+$('#genSDS').on('click', '[id*=dis-genSDS]', function () {
+	$("#sds_res").html('');
+});
+	
 //Generate ing sds
 $('#genSDS').on('click', '[id*=generateSDS]', function () {
 	$("#sds_res").html('<div class="alert alert-info"><img src="/img/loading.gif"/> Please wait, we generating your SDS</div>');