+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 6.3
 - Add phpMyAdmin for docker images
+- Add a message to notify user that SDS document generation its in preview state
 
 ### Version 6.2
 - Fix a bug giving an javascript error when you creating a new empty formula